% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adwords_completed_order_events.r
\name{adwords_completed_order_events}
\alias{adwords_completed_order_events}
\title{Retrieve Mixpanel AdWords Completed Order Events}
\usage{
adwords_completed_order_events(account, from = Sys.Date(), to = Sys.Date())
}
\arguments{
\item{account}{A mixpanel account object.}

\item{from}{Start date in either format <"yyyy-mm-dd"> or <yyyymmdd>. Inclusive.}

\item{to}{End date in either format <"yyyy-mm-dd"> or <yyyymmdd>. Inclusive.}
}
\description{
Wrapper function to get completed order events originating from AdWords
}
\examples{
adwords_completed_order_events(account, from=start_date, to=end_date)
}

