% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_ppc_completed_order_events.r
\name{all_ppc_completed_order_events}
\alias{all_ppc_completed_order_events}
\title{Retrieve Mixpanel Completed Order Events From All PPC Channels}
\usage{
all_ppc_completed_order_events(account = NULL, from = Sys.Date(),
  to = Sys.Date())
}
\arguments{
\item{account}{A mixpanel account object.}

\item{from}{Start date in either format <"yyyy-mm-dd"> or <yyyymmdd>. Inclusive.}

\item{to}{End date in either format <"yyyy-mm-dd"> or <yyyymmdd>. Inclusive.}
}
\value{
A list containing \code{start_time} (the start range of the events data), \code{end_time} (the end range of the events data) and \code{events} (the actual events).
}
\description{
Retrieves all ppc events (those with latest_ad_utm_source of "cpc") from Mixpanel.
If data with the start and end date specified has already been retrieved, it restores
the data from a local RData file, mixpanel_ppc_events.RData to save time.  
If data must be pulled, and an RMixpanel account object is not provided, the credentials
are looked up using a mixpanel_credentials.yml file in the working directory.  
mixpanel_credentials.yml must include four key value pairs: name, token, apiKey, apiSecret.
}
\examples{
bing_completed_order_events(account, from=start_date, to=end_date)
}

