% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bing_raw_completed_order_events.r
\name{bing_raw_completed_order_events}
\alias{bing_raw_completed_order_events}
\title{Retrieve Raw Mixpanel Bing Completed Order Events}
\usage{
bing_raw_completed_order_events(account = NULL, from = Sys.Date(),
  to = Sys.Date())
}
\arguments{
\item{account}{A mixpanel account object.}

\item{from}{Start date in either format <"yyyy-mm-dd"> or <yyyymmdd>. Inclusive.}

\item{to}{End date in either format <"yyyy-mm-dd"> or <yyyymmdd>. Inclusive.}
}
\description{
Wrapper function to get completed order events originating from Bing.  If an 
RMixpanel account object is not provided, the credentials are looked up using a
mixpanel_credentials.yml file in the working directory.  mixpanel_credentials.yml
must include four key value pairs: name, token, apiKey, apiSecret.
}
\examples{
bing_raw_completed_order_events(account, from=start_date, to=end_date)
}

