% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_adwords_raw_completed_order_events.r
\name{clean_adwords_raw_completed_order_events}
\alias{clean_adwords_raw_completed_order_events}
\title{Clean Adwords Raw Completed Order Eventss}
\usage{
clean_adwords_raw_completed_order_events(adwords_events)
}
\arguments{
\item{adwords_events}{A data frame of events specifically relating to AdWords from \code{all_ppc_raw_completed_order_events} or \code{adwords_raw_completed_order_events}}
}
\value{
A data frame with the events cleaned.
}
\description{
Cleans the data from Mixpanel related to AdWords completed order events.
}
\examples{
clean_adwords_raw_completed_order_events(adwords_events)
}

