% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_completed_order_events.r
\name{clean_completed_order_events}
\alias{clean_completed_order_events}
\title{Clean Raw Completed Order Eventss}
\usage{
clean_completed_order_events(events_matrix)
}
\arguments{
\item{events_data_frame}{A matrix of events specifically relating to ppc events from Mixpanel generated by \code{all_ppc_raw_completed_order_events}, \code{adwords_raw_completed_order_events}, \code{bing_raw_completed_order_events}}
}
\value{
A data frame with the events cleaned.
}
\description{
Cleans the Completed Order events data from Mixpanel.
}
\examples{
clean_completed_order_events(adwords_events)
}

